[integ-tests] Add p6-b200 to test_efa
p6-b200 requires a capacity block. This commit makes the test to automatically find the capacity block by reusing the same logic as test_gb200. This requires a little bit of refactor

Using capacity block requires using RunInstances instead of CreateFleet. Therefore, this commit changes the cluster configuration file of test_efa to always use RunInstances
