+++ tests/integration-tests/configs/develop.yaml
@@ -264,6 +264,10 @@ test-suites:
           instances: ["p4d.24xlarge"]
           oss: [{{ NO_RHEL_OS_X86_1 }}]  # The capacity reservation cannot use RHEL operating system
           schedulers: ["slurm"]
+        - regions: [ "use1-az6" ]  # do not move, unless capacity reservation is moved as well
+          instances: [ "p6-b200.48xlarge" ]
+          oss: ["alinux2023", "ubuntu2204", "ubuntu2404", "rocky8", "rocky9"]
+          schedulers: [ "slurm" ]
         - regions: [{{ c6gn_16xlarge_CAPACITY_RESERVATION_2_INSTANCES_2_HOURS_YESPG_OS_ARM_0 }}]
           instances: ["c6gn.16xlarge"]
           oss: [{{ OS_ARM_0 }}]
