+++ tests/integration-tests/tests/common/utils.py
@@ -552,3 +552,22 @@ def terminate_nodes_manually(instance_ids, region):
         assert_that(instance_states.get("InstanceId")).is_equal_to(instance_id)
         assert_that(instance_states.get("CurrentState").get("Name")).is_in("shutting-down", "terminated")
     logging.info("Terminated nodes: {}".format(instance_ids))
+
+
+def get_capacity_reservation_id(instance_type, region, count):
+    ec2_client = boto3.client("ec2", region_name=region)
+    paginator = ec2_client.get_paginator("describe_capacity_reservations")
+    # List to store matching reservation IDs
+    reservations_ids = []
+    # Paginate through the results
+    for page in paginator.paginate():
+        for reservation in page.get("CapacityReservations", []):
+            if instance_type == reservation.get("InstanceType") and reservation.get("AvailableInstanceCount") >= count:
+                reservations_ids.append(
+                    {
+                        "CapacityReservationId": reservation["CapacityReservationId"],
+                        "TotalInstanceCount": reservation["TotalInstanceCount"],
+                        "AvailableInstanceCount": reservation["AvailableInstanceCount"],
+                    }
+                )
+    return reservations_ids
